ENGR00138491-1: fix WiFi deep sleep mode blocks the system to suspend issue
When the WiFi is entering deepsleep for suspend, it will acquire a
wakelock with 3 seconds timeout. However, to system suspend operation,
3 seconds is too long and then obviously will l block the system to
suspend successfully.
That why we always failed to suspend and just resumed directly and then
falled into a dead loop that suspend/resume repeatly.

Changing to not acquire the wakelock for only disconnect event.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>